+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## 0.2.6 2023-10
+## 0.2.7 2023-10
+
+## 0.2.6 2023-10-24
 
 * [Deprecated `loaded-protyle` use `loaded-protyle-static` instead](https://github.com/siyuan-note/siyuan/issues/9468)